Module sui_verifier_v2::id_leak_verifier

Objects whose struct type has key ability represent Sui objects. They have unique IDs that should never be reused. This verifier makes sure that the id field of Sui objects never get leaked. Unpack is the only bytecode that could extract the id field out of a Sui object. From there, we track the flow of the value and make sure it can never get leaked outside of the function. There are four ways it can happen:

Returned

Written into a mutable reference

Added to a vector

Passed to a function cal::;

Functions

verify_module